增加一个类型参数 增加新的依赖项

+++ include/circos/shapes/line.h
@@ -27,7 +27,7 @@ namespace spiritsaway::circos
 		}
 		float len() const
 		{
-			return std::sqrtf(len_square());
+			return std::sqrtf(static_cast<float>(len_square()));
 		}
 		int len_square() const
 		{
